+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTTH7ER8SR9XRZP532XFGA13"
+        "01KTV1B07HB1NKBWXRMG0Z1M0M"
     )
     r"""The unique identifier of the data source"""