+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2057,7 +2057,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KTTH7EF9APFANT61STX020Q9"
+    id: Optional[str] = "01KTV1AZXY7T49X6VWXKJCVQW9"
 
     description: Optional[str] = None