+++ packages/orq-rc/RELEASES.md
@@ -8629,3 +8629,13 @@ Based on:
 - [python v4.11.0-rc.12] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.12] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.12 - packages/orq-rc
+
+## 2026-06-11 09:51:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.776.0 (2.900.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.13] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.13 - packages/orq-rc